renamed userFacing to userFacingStrings
