+++ src/utils/formatting/stringFormatting.js
@@ -1,7 +1,7 @@
 import numeral from 'numeral';
 
 import { NUMBER_FORMAT_DEFAULT } from '../../constants/utilityConstants';
-import { BLANK_FIELD } from '../../constants/userFacing';
+import { BLANK_FIELD } from '../../constants/userFacingStrings';
 import { getSelectedStockValueForKey } from '../stateGetters';
 
 const formatValueForDisplay = (value, formatSpecifier) => {
